Test: DOM elements clearing
