+++ src/index.js
@@ -7,16 +7,11 @@ import { setLocalStorage, updateLocalStorage } from './modules/local_storage.js'
 import {
   addTask, editDescription, deleteTask, updateTaskIndex, toogleBoolean, deleteCompleted,
 } from './modules/crud.js';
+import cleanDisplay from './modules/cleanDisplay.js';
 
 const toDoListCointainer = document.getElementById('toDoListContainer');
 const newTaskInput = document.getElementById('newTaskInput');
 const toDoList = [];
-
-function cleanDisplay(arr) {
-  while (arr.firstChild) {
-    arr.removeChild(arr.firstChild);
-  }
-}
 
 function displayElement(arr) {
   cleanDisplay(toDoListCointainer);
