+++ src/modules/cleanDisplay.js
@@ -1,0 +1,8 @@
+function cleanDisplay(arr) {
+  while (arr.firstChild) {
+    arr.removeChild(arr.firstChild);
+  }
+  return arr;
+}
+
+export default cleanDisplay;
